resolved conflicts

+++ scylla-persistence/src/main/java/com/netflix/conductor/scylla/dao/ScyllaExecutionDAO.java
@@ -517,7 +517,7 @@ public class ScyllaExecutionDAO extends ScyllaBaseDAO
     }
 
     private TaskModel getTaskModel(String shardId, String taskId, String workflowId) {
-        Integer correlationId = Objects.isNull(shardId) ? 0 : Integer.parseInt(shardId);
+        Long correlationId = Objects.isNull(shardId) ? 0L : Long.parseLong(shardId);
         if (workflowId == null) {
             return null;
         }
@@ -727,7 +727,7 @@ public class ScyllaExecutionDAO extends ScyllaBaseDAO
     @Override
     public WorkflowModel getWorkflow(String shardId, String workflowId, boolean includeTasks) {
         UUID workflowUUID = toUUID(workflowId, "Invalid workflow id");
-        Integer correlationId = Objects.isNull(shardId) ? 0 : Integer.parseInt(shardId);
+        Long correlationId = Objects.isNull(shardId) ? 0L : Long.parseLong(shardId);
 
         return getWorkflowModel(workflowId, includeTasks, workflowUUID, correlationId);
     }
@@ -741,7 +741,7 @@ public class ScyllaExecutionDAO extends ScyllaBaseDAO
         return getWorkflowModel(workflowId, includeTasks, workflowUUID, correlationId);
     }
 
-    private WorkflowModel getWorkflowModel(String workflowId, boolean includeTasks, UUID workflowUUID, Integer correlationId) {
+    private WorkflowModel getWorkflowModel(String workflowId, boolean includeTasks, UUID workflowUUID, Long correlationId) {
         try {
             WorkflowModel workflow = null;
             ResultSet resultSet;
